split: extract _expressions.py, _capture.py, _outcome.py from fleet/_api.py
Extract three cohesion-based modules from the 1056-line fleet/_api.py:

_expressions.py (~170 LOC): evaluate_skip_when, _interpolate_campaign_refs,
  _CAMPAIGN_REF_RE, _INPUTS_REF_RE, _strip_quotes

_capture.py (~160 LOC): _extract_captures, _normalize_capture_spec,
  CaptureCompletenessError, _validate_capture_value

_outcome.py (~70 LOC): classify_dispatch_outcome, _is_abandon_reason,
  _checkpoint_to_dict

Rump _api.py (~756 LOC) retains execute_dispatch, _run_dispatch, and
lifecycle helpers. __init__.py re-exports updated to new modules.

Structural guard test added (test_api_split_integrity.py). Arch test
exemptions updated (fleet file ceiling 16→19, _api.py line exemption
removed, _outcome.py added to StrEnum/precautionary skip paths). Test
import paths updated for affected fleet tests.
