+++ src/autoskillit/fleet/__init__.py
@@ -4,20 +4,18 @@ Gateway exports per REQ-IMP-001 — consumers import from
 ``autoskillit.fleet``, not from sub-modules.
 """
 
-from ._api import (
-    CaptureCompletenessError,
-    classify_dispatch_outcome,
-    evaluate_skip_when,
-    execute_dispatch,
-)
 from ._api import _write_pid as _write_pid
+from ._api import execute_dispatch
+from ._capture import CaptureCompletenessError
 from ._checkpoint_bridge import checkpoint_from_sidecar
+from ._expressions import evaluate_skip_when
 from ._label_cleanup import (
     cleanup_orphaned_labels,
     discover_campaign_state_files,
     sweep_stale_dispatch_labels,
 )
 from ._liveness import is_dispatch_session_alive
+from ._outcome import classify_dispatch_outcome
 from ._prompts import _build_admiral_dispatch_block as _build_admiral_dispatch_block
 from ._prompts import _build_food_truck_prompt as _build_food_truck_prompt
 from ._semaphore import FleetSemaphore
