+++ tests/fleet/CLAUDE.md
@@ -10,6 +10,7 @@ Fleet campaign dispatch, state persistence, and sidecar tests.
 | `_helpers.py` | Shared helpers for tests/fleet/ test modules |
 | `conftest.py` | Shared fixtures for tests/fleet/ |
 | `test_api.py` | Tests for fleet._api module (Group J) |
+| `test_api_split_integrity.py` | Structural guard: fleet `_api.py` split — verifies new modules export expected symbols and public API surface is preserved |
 | `test_api_dispatch_marker.py` | Tests for _run_dispatch marker lifecycle and _touch_dispatch_marker heartbeat |
 | `test_campaign_capture.py` | Tests for campaign capture extraction and ingredient interpolation (Group J) |
 | `test_capture_roundtrip.py` | Tests for prompt-extractor field name alignment — verifies sentinel examples use bare names matching `_extract_captures` expectations |
